+++ src/main/java/de/maxhenkel/corpse/entities/CorpseRenderState.java
@@ -3,6 +3,7 @@ package de.maxhenkel.corpse.entities;
 import net.minecraft.client.renderer.entity.state.AvatarRenderState;
 import net.minecraft.client.renderer.entity.state.EntityRenderState;
 import net.minecraft.client.renderer.entity.state.SkeletonRenderState;
+import net.minecraft.world.entity.EntityType;
 
 public class CorpseRenderState extends EntityRenderState {
 
@@ -10,5 +11,9 @@ public class CorpseRenderState extends EntityRenderState {
     public AvatarRenderState playerRenderState = new AvatarRenderState();
     public SkeletonRenderState skeletonRenderState = new SkeletonRenderState();
     public boolean skeleton;
+
+    public CorpseRenderState() {
+        skeletonRenderState.entityType = EntityType.SKELETON;
+    }
 
 }